// src/database/entities/expense-policy.enums.ts
export enum ExpensePolicyRuleType {
  CONDITIONAL = 'CONDITIONAL',
  RECEIPT_REQUIRED = 'RECEIPT_REQUIRED',
  CATEGORY_MONTHLY_CAP = 'CATEGORY_MONTHLY_CAP',
  WEEKEND_TRAVEL_JUSTIFICATION = 'WEEKEND_TRAVEL_JUSTIFICATION',
  DUPLICATE_DETECTION = 'DUPLICATE_DETECTION',
}

export enum ExpensePolicySeverity {
  BLOCK = 'BLOCK',
  WARN = 'WARN',
}